Changed the type of vlan id from 16 bit to 32 bit
Vlan interface is having vlanid as 32 bit but the create
interface was having 16 bit so this commit changes the
type of vlanid property in the create interface to 32 bit.

+++ network_manager.cpp
@@ -83,7 +83,7 @@ void Manager::createChildObjects()
 
 }
 
-void Manager::vLAN(IntfName interfaceName, uint16_t id)
+void Manager::vLAN(IntfName interfaceName, uint32_t id)
 {
 }
